// src/app/Model/Configuration/payeridentifier.ts
export class Payeridentifier {
  id: string;
  sid: string;
  status: string;
  userid: string;
  payerid: string;
  username: string;
  createdon: string;
  payername: string;
  separator: string;
  identifier_1: string;
  identifier_2: string;
  identifier_3: string;
  identifier_4: string;
  identifier_5: string;
  adjustmentreasoncode: string;
}